refactor(workflow): split platform workflow classes
Why:

- Make the platform workflow service easier to read by moving workflow classes out of the monolithic workflows module.

What:

- Move init, mutating, destructive, verify, and step protocol classes into focused workflow modules.

- Keep workflow.workflows as a compatibility facade for existing imports.

How:

- Preserve existing class names, public exports, and shared runtime helper behavior while relocating code by responsibility.

- Keep the change inside the application service workflow package without adding infrastructure dependencies.

Verification:

- PYTHONPATH=src python3 -m unittest tests.application.services.platform.test_platform_workflows tests.application.services.platform.test_platform_service_exports: pass

- PYTHONPATH=src python3 -m unittest tests.infrastructure.test_composition tests.architecture.test_installation_method_trace_coverage: pass

- python3 tools/quality_gate.py test: pass, 1051 tests, 28 skipped

- python3 -m compileall -q src/tiny_swarm_world/application/services/platform/workflow: pass

- git diff --cached --check: pass

Impact:

- Affects platform workflow application-service module organization only; existing import facade remains compatible.

- Breaking changes: None.

Limitations:

- Full quality gate was not run because WSL lacks ruff and mypy; test gate passed.
